Move Java stuff to JavaPluginDescription.
